requirements.txt
tensorflow-cpu==2.3.1
# https://download.pytorch.org/whl/cpu/torch-1.7.0%2Bcpu-cp38-cp38-linux_x86_64.whl
tf-nightly-cpu
requests==2.25.0
Pillow==8.0.1
matplotlib==3.3.3
pycocotools==2.0.2
numpy<1.19.0
flask==1.1.2
gunicorn==20.0.4
protobuf==3.14.0
